+++ packages/react-native/src/evm/components/ConnectWalletDetails/WalletDetailsButton.tsx
@@ -10,7 +10,7 @@ import { useMemo, useState } from "react";
 import { TextBalance } from "../base/TextBalance";
 import { SupportedTokens } from "../SendFunds/defaultTokens";
 import { SMART_WALLET_ICON } from "../../assets/svgs";
-import { ImageSvgUri } from "../base";
+import { WalletIcon } from "../base";
 import { useLocale } from "../../providers/ui-context-provider";
 
 export type ConnectWalletDetailsProps = {
@@ -92,13 +92,7 @@ export const WalletDetailsButton = ({
             alignContent="center"
             justifyContent="flex-start"
           >
-            <Box borderRadius="lg" overflow="hidden">
-              <ImageSvgUri
-                width={32}
-                height={32}
-                imageUrl={avatarUrl || walletIconUrl}
-              />
-            </Box>
+            <WalletIcon iconUri={avatarUrl || walletIconUrl} size={32} />
             <Box ml="md" justifyContent="center" alignItems="flex-start">
               {activeWallet?.walletId === LocalWallet.id ? (
                 <Text variant="bodySmall" color="red">