+++ packages/react-native/src/evm/wallets/connectors/embedded-wallet/embedded/auth.ts
@@ -1,14 +1,15 @@
+import type { CognitoUser } from "amazon-cognito-identity-js";
+
 import {
   AuthProvider,
   AuthStoredTokenWithCookieReturnType,
-  RecoveryShareManagement,
-} from "@paperxyz/embedded-wallet-service-sdk";
-import type { CognitoUser } from "amazon-cognito-identity-js";
-import { Auth } from "aws-amplify";
-
+  SendEmailOtpReturnType,
+} from "@thirdweb-dev/wallets";
 import {
   generateAuthTokenFromCognitoEmailOtp,
   getEmbeddedWalletUserDetail,
+  sendUserManagedEmailOtp,
+  validateUserManagedEmailOtp,
   verifyClientId,
 } from "./helpers/api/fetchers";
 import {
@@ -20,17 +21,17 @@ import {
   postPaperAuthUserManaged,
   prePaperAuth,
 } from "./helpers/auth/middleware";
-import { isDeviceSharePresentForUser } from "./helpers/storage/local";
 import { getCognitoUser, setCognitoUser } from "./helpers/storage/state";
-import { SendEmailOtpReturnType } from "@thirdweb-dev/wallets";
-import { InAppBrowser } from "react-native-inappbrowser-reborn";
-import { AuthOptions, OauthOption } from "../types";
+import { isDeviceSharePresentForUser } from "./helpers/storage/local";
+import { Auth } from "aws-amplify";
 import {
   ROUTE_AUTH_JWT_CALLBACK,
   ROUTE_HEADLESS_GOOGLE_LOGIN,
 } from "./helpers/constants";
+import { AuthOptions, OauthOption, VerifiedTokenResponse } from "../types";
+import { InAppBrowser } from "react-native-inappbrowser-reborn";
 
-export async function sendEmailOTP(options: {
+export async function sendVerificationEmail(options: {
   email: string;
   clientId: string;
 }): Promise<SendEmailOtpReturnType> {
@@ -40,16 +41,6 @@ export async function sendEmailOTP(options: {
     authenticationMethod: AuthProvider.COGNITO,
     email: options.email,
   });
-
-  // AWS Auth flow
-  let cognitoUser: CognitoUser;
-  try {
-    cognitoUser = await cognitoEmailSignIn(options.email, options.clientId);
-  } catch (e) {
-    await cognitoEmailSignUp(options.email, options.clientId);
-    cognitoUser = await cognitoEmailSignIn(options.email, options.clientId);
-  }
-  setCognitoUser(cognitoUser);
 
   let result: Awaited<ReturnType<typeof getEmbeddedWalletUserDetail>>;
   try {
@@ -63,11 +54,29 @@ export async function sendEmailOTP(options: {
     );
   }
 
+  if (result.recoveryShareManagement === "USER_MANAGED") {
+    try {
+      await sendUserManagedEmailOtp(options.email, options.clientId);
+    } catch (error) {
+      throw new Error(`Error sending user managed email otp: ${error}`);
+    }
+  } else {
+    // CLOUD_MANAGED
+    let cognitoUser: CognitoUser;
+    try {
+      cognitoUser = await cognitoEmailSignIn(options.email, options.clientId);
+    } catch (e) {
+      await cognitoEmailSignUp(options.email, options.clientId);
+      cognitoUser = await cognitoEmailSignIn(options.email, options.clientId);
+    }
+    setCognitoUser(cognitoUser);
+  }
+
   return result.isNewUser
     ? {
         isNewUser: result.isNewUser,
         isNewDevice: true,
-        recoveryShareManagement: RecoveryShareManagement.AWS_MANAGED,
+        recoveryShareManagement: result.recoveryShareManagement,
       }
     : {
         isNewUser: result.isNewUser,
@@ -75,60 +84,88 @@ export async function sendEmailOTP(options: {
           options.clientId,
           result.walletUserId ?? "",
         )),
-        recoveryShareManagement: RecoveryShareManagement.AWS_MANAGED,
+        recoveryShareManagement: result.recoveryShareManagement,
       };
 }
 
-export async function validateEmailOTP({
-  clientId,
-  otp,
-}: {
+export async function validateEmailOTP(options: {
+  email: string;
   otp: string;
   clientId: string;
+  recoveryCode?: string;
 }): Promise<AuthStoredTokenWithCookieReturnType> {
-  let verifiedToken: Awaited<
-    ReturnType<typeof generateAuthTokenFromCognitoEmailOtp>
-  >["verifiedToken"];
-  let verifiedTokenJwtString: string;
-
+  let result: Awaited<ReturnType<typeof getEmbeddedWalletUserDetail>>;
   try {
-    let cognitoUser = getCognitoUser();
-    if (!cognitoUser) {
-      throw new Error("MISSING COGNITO USER");
-    }
-    cognitoUser = await Auth.sendCustomChallengeAnswer(cognitoUser, otp);
-
-    // It we get here, the answer was sent successfully,
-    // but it might have been wrong (1st or 2nd time)
-    // So we should test if the user is authenticated now
-    const session = await Auth.currentSession();
-
-    ({ verifiedToken, verifiedTokenJwtString } =
-      await generateAuthTokenFromCognitoEmailOtp(session, clientId));
+    result = await getEmbeddedWalletUserDetail({
+      email: options.email,
+      clientId: options.clientId,
+    });
   } catch (e) {
-    throw new Error(`Invalid OTP ${e}`);
+    throw new Error(
+      `Malformed response from the send email OTP API: ${JSON.stringify(e)}`,
+    );
+  }
+  let verifiedTokenResponse: VerifiedTokenResponse;
+
+  if (result.recoveryShareManagement === "USER_MANAGED") {
+    try {
+      verifiedTokenResponse = await validateUserManagedEmailOtp({
+        email: options.email,
+        otp: options.otp,
+        clientId: options.clientId,
+      });
+    } catch (error) {
+      throw new Error(`Error validating user managed email otp: ${error}`);
+    }
+  } else {
+    try {
+      let cognitoUser = getCognitoUser();
+      if (!cognitoUser) {
+        throw new Error("MISSING COGNITO USER");
+      }
+      cognitoUser = await Auth.sendCustomChallengeAnswer(
+        cognitoUser,
+        options.otp,
+      );
+
+      // It we get here, the answer was sent successfully,
+      // but it might have been wrong (1st or 2nd time)
+      // So we should test if the user is authenticated now
+      const session = await Auth.currentSession();
+
+      verifiedTokenResponse = await generateAuthTokenFromCognitoEmailOtp(
+        session,
+        options.clientId,
+      );
+    } catch (e) {
+      throw new Error(`Invalid OTP ${e}`);
+    }
   }
 
   try {
     const storedToken: AuthStoredTokenWithCookieReturnType["storedToken"] = {
-      jwtToken: verifiedToken.jwtToken,
-      authDetails: verifiedToken.authDetails,
-      authProvider: verifiedToken.authProvider,
-      developerClientId: verifiedToken.developerClientId,
-      cookieString: verifiedTokenJwtString,
+      jwtToken: verifiedTokenResponse.verifiedToken.jwtToken,
+      authDetails: verifiedTokenResponse.verifiedToken.authDetails,
+      authProvider: verifiedTokenResponse.verifiedToken.authProvider,
+      developerClientId: verifiedTokenResponse.verifiedToken.developerClientId,
+      cookieString: verifiedTokenResponse.verifiedTokenJwtString,
       // we should always store the jwt cookie since there's no concept of cookie in react native
       shouldStoreCookieString: true,
-      isNewUser: verifiedToken.isNewUser,
+      isNewUser: verifiedTokenResponse.verifiedToken.isNewUser,
     };
 
-    await postPaperAuth(storedToken, clientId);
+    await postPaperAuth({
+      storedToken,
+      clientId: options.clientId,
+      recoveryCode: options.recoveryCode,
+    });
 
     return { storedToken };
   } catch (e) {
     throw new Error(
-      `Malformed response from the verify one time password: ${JSON.stringify(
-        e,
-      )}`,
+      `Malformed response from the verify one time password: ${
+        (e as Error).message
+      }}`,
     );
   }
 }
@@ -195,7 +232,7 @@ export async function socialLogin(oauthOptions: OauthOption, clientId: string) {
       isNewUser: storedToken.isNewUser,
     };
 
-    await postPaperAuth(toStoreToken, clientId);
+    await postPaperAuth({ storedToken: toStoreToken, clientId });
 
     return { storedToken, email: storedToken.authDetails.email };
   } catch (e) {
@@ -218,7 +255,7 @@ export async function customJwt(authOptions: AuthOptions, clientId: string) {
   });
   if (!resp.ok) {
     const error = await resp.json();
-    throw new Error(`JWT authentication error: ${error.message} `);
+    throw new Error(`JWT authentication error: ${error.message}`);
   }
 
   try {