+++ packages/sdk/src/evm/common/nft.ts
@@ -14,9 +14,6 @@ import type {
   IERC165,
   IERC721Metadata,
 } from "@thirdweb-dev/contracts-js";
-import ERC165MetadataAbi from "@thirdweb-dev/contracts-js/dist/abis/IERC165.json";
-import ERC721MetadataAbi from "@thirdweb-dev/contracts-js/dist/abis/IERC721Metadata.json";
-import ERC1155MetadataAbi from "@thirdweb-dev/contracts-js/dist/abis/IERC1155Metadata.json";
 import type {
   ThirdwebStorage,
   UploadProgressEvent,
@@ -109,6 +106,9 @@ export async function fetchTokenMetadataForContract(
   storage: ThirdwebStorage,
 ): Promise<NFTMetadata> {
   let uri: string | undefined;
+  const ERC165MetadataAbi = (
+    await import("@thirdweb-dev/contracts-js/dist/abis/IERC165.json")
+  ).default;
   const erc165 = new Contract(
     contractAddress,
     ERC165MetadataAbi,
@@ -119,6 +119,9 @@ export async function fetchTokenMetadataForContract(
     erc165.supportsInterface(InterfaceId_IERC1155),
   ]);
   if (isERC721) {
+    const ERC721MetadataAbi = (
+      await import("@thirdweb-dev/contracts-js/dist/abis/IERC721Metadata.json")
+    ).default;
     const erc721 = new Contract(
       contractAddress,
       ERC721MetadataAbi,
@@ -126,6 +129,9 @@ export async function fetchTokenMetadataForContract(
     ) as IERC721Metadata;
     uri = await erc721.tokenURI(tokenId);
   } else if (isERC1155) {
+    const ERC1155MetadataAbi = (
+      await import("@thirdweb-dev/contracts-js/dist/abis/IERC1155Metadata.json")
+    ).default;
     const erc1155 = new Contract(
       contractAddress,
       ERC1155MetadataAbi,